start work on extracting mobile nav bar
