+++ src/routes/__layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Footer from "../components/Footer.svelte";
-  import NavBar from "../components/NavBar.svelte";
+  import NavBar from "../components/NavBars/NavBar.svelte";
   import Seo from "../components/Seo.svelte";
   import Blob from "../components/Blobs/Blob.svelte";
   import BlobBackground from "../components/Blobs/BlobBackground.svelte";
